build: fix 'Unmet peer dependencies' (#879)
fix #878
