+++ packages/components/_private/selector/src/Selector.tsx
@@ -21,8 +21,8 @@ import { isString } from 'lodash-es'
 import { callEmit, offResize, onResize } from '@idux/cdk/utils'
 import { ɵOverflow } from '@idux/components/_private/overflow'
 import { useGlobalConfig } from '@idux/components/config'
+import { useFormSize } from '@idux/components/form'
 import { IxIcon } from '@idux/components/icon'
-import { useFormSize } from '@idux/components/utils'
 
 import { useInputState } from './composables/useInputState'
 import Input from './contents/Input'
