+++ packages/components/cascader/src/Cascader.tsx
@@ -11,8 +11,9 @@ import { type VKey, useState } from '@idux/cdk/utils'
 import { ɵOverlay } from '@idux/components/_private/overlay'
 import { ɵSelector, type ɵSelectorInstance } from '@idux/components/_private/selector'
 import { useGlobalConfig } from '@idux/components/config'
+import { useFormAccessor } from '@idux/components/form'
 import { ɵUseOverlayState } from '@idux/components/select'
-import { useFormAccessor, useGetKey } from '@idux/components/utils'
+import { useGetKey } from '@idux/components/utils'
 
 import { useActiveState } from './composables/useActiveState'
 import { useDataSource } from './composables/useDataSource'
