+++ packages/components/input-number/src/InputNumber.tsx
@@ -11,8 +11,8 @@ import { computed, defineComponent, normalizeClass, onMounted, ref } from 'vue'
 
 import { ɵInput } from '@idux/components/_private/input'
 import { useGlobalConfig } from '@idux/components/config'
+import { useFormFocusMonitor, useFormSize } from '@idux/components/form'
 import { IxIcon } from '@idux/components/icon'
-import { useFormFocusMonitor, useFormSize } from '@idux/components/utils'
 
 import { inputNumberProps } from './types'
 import { useInputNumber } from './useInputNumber'
