+++ packages/components/package.json
@@ -36,14 +36,15 @@
   "dependencies": {
     "@idux/cdk": "^1.0.0-beta.7",
     "date-fns": "^2.27.0",
-    "lodash-es": "^4.17.21",
+    "lodash-es": "^4.17.0",
     "vue-types": "^3.0.0"
   },
   "devDependencies": {
     "axios": "^0.25.0",
+    "vue": "^3.2.0",
     "vue-router": "^4.0.12"
   },
   "peerDependencies": {
-    "vue": "^3.2.29"
+    "vue": "^3.2.0"
   }
 }
