+++ packages/components/utils/index.ts
@@ -8,5 +8,4 @@
 export * from './src/colors'
 export * from './src/convertTarget'
 export * from './src/convertVNode'
-export * from './src/form'
 export * from './src/useKey'
